:sparkles :适配了mv的点击事件，item旁边的mv图标也可以点击了

+++ module_find/src/main/java/com/handsome/module/find/network/model/AlbumData.kt
@@ -112,7 +112,7 @@ data class AlbumData(
         val l: L,
         val m: M,
         val mst: Int,
-        val mv: Int,
+        val mv: Long,
         val name: String,
         val no: Int,
         val noCopyrightRcmd: Any,
